Clarification: Pages branch has to be the branch of a public repository (#204)
The [pages manual](https://codeberg.page/) explicitly mentions that a public repository has to be created for the pages repository. That is not mentioned for the pages branch, consequently one could assume that this can be a private repository as well - which is not the case.

I hereby propose to add the word *public* to the manual where the branch is mentioned in order to clarify it and prevent the described confusion.

Co-authored-by: Cs137 <cs137@mailbox.org>
Reviewed-on: https://codeberg.org/Codeberg/pages-server/pulls/204
Reviewed-by: 6543 <6543@obermui.de>
Co-authored-by: Cs137 <cs137@noreply.codeberg.org>
Co-committed-by: Cs137 <cs137@noreply.codeberg.org>